+++ test/test_helper.rb
@@ -1,6 +1,0 @@
-require "rubygems"
-require "cutest"
-
-$VERBOSE = 1
-
-require File.join(File.dirname(__FILE__), "..", "lib", "ost")